窗口布局

+++ src/utils/src/terminal.cpp
@@ -1,20 +1,20 @@
-#include <terminal/terminal.h>
+#include <utils/terminal.h>
 
 namespace tetriz
 {
 inline constexpr std::string_view CSI = "\033[";
 
-void move_to(int row, int col)
+void move_to(size_t row, size_t col)
 {
 	fmt::print("{}{};{}H", CSI, row, col);
 }
 
-void set_fore_color(int color_id)
+void set_fore_color(size_t color_id)
 {
 	fmt::print("{}38;5;{}m", CSI, color_id);
 }
 
-void set_back_color(int color_id)
+void set_back_color(size_t color_id)
 {
 	fmt::print("{}48;5;{}m", CSI, color_id);
 }
